Update app.py endpoints for user authentication

+++ app.py
@@ -7,7 +7,9 @@ from flask import Flask
 from flask import request
 import requests
 from datetime import timedelta
+from datetime import datetime
 import os
+import users_dao
 
 
 app = Flask(__name__)
@@ -26,6 +28,19 @@ def success_response(data, code=200):
 
 def failure_response(data, code=404):
     return json.dumps(data), code
+
+def extract_token(request):
+    """
+    Helper function that extracts the token from the header of a request
+    """
+    auth_header = request.headers.get("Authorization")
+
+    if auth_header is None:
+        return False, json.dumps({"Missing authorization header"})
+
+    bearer_token = auth_header.replace("Bearer", "").strip()
+
+    return True, bearer_token
 
 
 @app.route("/")
@@ -47,21 +62,101 @@ def get_user(user_id):
         return failure_response({"error": "user not found"})
     return success_response(user.serialize())
 
-    
+
 @app.route("/api/user/", methods = ["POST"])
 def create_user():
     """
-    Creates a user
+    Endpoint for registering a new user
+    """
+    body = json.loads(request.body)
+    email = body.get("email")
+    password = body.get("password")
+
+    if email is None or password is None:
+        return failure_response({"Error": "Bad Request"}, 400)
+
+    was_successful, user = users_dao.create_user(email, password)
+
+    if not was_successful:
+        return failure_response("User email already exists.")
+
+    return success_response(
+        {
+            "session_token": user.session_token,
+            "session_expiration": str(user.session_expiration),
+            "update_token": user.update_token
+        }, 201
+    )
+
+
+@app.route("/api/users/login/", methods=["POST"])
+def login():
+    """
+    Endpoint for logging in a user.
     """
     body = json.loads(request.data)
-    username = body.get("username")
+    email = body.get("email")
     password = body.get("password")
-    if not username or not password:
-        return failure_response({"Error": "Bad Request"}, 400)
-    new_user = User(username=username, password=password)
-    db.session.add(new_user)
+
+    if email is None or password is None:
+        return failure_response("Missing email or password")
+
+    was_successful, user = users_dao.verify_credentials(email, password)
+
+    if not was_successful:
+        return failure_response("Incorrect username or password.")
+
+    return success_response(
+        {
+            "session_token": user.session_token,
+            "session_expiration": str(user.session_expiration),
+            "update_token": user.update_token
+        }, 201
+    )
+
+
+@app.route("/api/users/session/", methods=["POST"])
+def update_session():
+    """
+    Endpoint for updating a user's session
+    """
+    was_successful, update_token = extract_token(request)
+
+    if not was_successful:
+        return update_token
+
+    try:
+        user = users_dao.renew_session(update_token)
+    except Exception as e:
+        return failure_response(f"Invalid update token: {str(e)}")
+
+    return success_response(
+        {
+            "session_token": user.session_token,
+            "session_expiration": str(user.session_expiration),
+            "update_token": user.update_token
+        }, 201
+    )
+
+
+@app.route("/api/users/logout/", methods=["POST"])
+def logout():
+    was_successful, session_token = extract_token(request)
+
+    if not was_successful:
+        return session_token
+
+    user = users_dao.get_user_by_session_token(session_token)
+
+    if not user or not user.verify_session_token(session_token):
+        return failure_response("Invalid session token")
+
+    user.session_expiration = datetime.now()
     db.session.commit()
-    return success_response(new_user.serialize(), 201)
+
+    return success_response({
+        "message": "User Logged out."
+    })
 
 
 @app.route("/api/user/<int:user_id>/", methods = ["POST"])
@@ -69,6 +164,15 @@ def create_event(user_id):
     """
     Create an event for a user
     """
+    was_successful, session_token = extract_token(request)
+
+    if not was_successful:
+        return session_token
+
+    user = users_dao.get_user_by_session_token(session_token)
+    if not user or not user.verify_session_token(session_token):
+        return failure_response("Invalid session token")
+
     body = json.loads(request.data)
     name = body.get("name")
     datetime = body.get("datetime")
@@ -77,10 +181,8 @@ def create_event(user_id):
     arrival = body.get("arrival")
     if not name or not datetime or not duration or not location or not arrival:
         return failure_response({"Error": "Bad Request"}, 400)
-    user = User.query.filter_by(id = user_id).first()
-    if not user:
-         return failure_response({"error": "user not found"})
-    new_event = Event(name=name, datetime=datetime, duration=duration, location=location, arrival=arrival, user=user_id)
+
+    new_event = Event(name=name, datetime=datetime, duration=duration, location=location, arrival=arrival, user=user.id)
     user.events.append(new_event)
     db.session.add(new_event)
     db.session.commit()
@@ -92,9 +194,22 @@ def delete_event(event_id):
     """
     delete event by id
     """
+    was_successful, session_token = extract_token(request)
+
+    if not was_successful:
+        return session_token
+
+    user = users_dao.get_user_by_session_token(session_token)
+    if not user or not user.verify_session_token(session_token):
+        return failure_response("Invalid session token")
+
     event = Event.query.filter_by(id = event_id).first()
     if not event:
         return failure_response("event not found")
+
+    if event.users != user.id:
+        return failure_response("Event does not belong to the user.")
+
     db.session.delete(event)
     db.session.commit()
     return success_response(event.serialize())
